telegram-bot/requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
urllib3>=2.0.7
httpx>=0.25.1
pytz>=2023.3
